tests/wpt/tests/dom/nodes/Document-contentType/contentType/contenttype_mimeheader_01.html

<!DOCTYPE html>
<title>Custom document.contentType === 'text/xml' when explicitly set to this value</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement('iframe');
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "text/xml");
  }), false);
  iframe.src = "../support/contenttype_setter.py?type=text&subtype=xml";
  document.body.appendChild(iframe);
});
</script>
